refactor(mergeToConference()): refine merging flag setting
Because the concurrency behaviour of the server, we cannot sure the merging process is over when the
function's procedure has finshed. But becasue the webphone._onCallEndFunc() is called whenever a
session is terminated, then we can hook the setting process of the `isMerging` flag to the
`terminated` event of the sessions using to merge.

RCINT-7614