fix(ConferenceCall): fix infinite spinner
Because there's a chance that the session closed before the _mergeToConference() finishing, so we
need to add the fucntion finishing signal to the door of sip instances' terminations. And also we
have discovered that if we close the session initiatively after the bring-in, the session will still
remain in the conference, so we do so to prevent the remaining session issue when duplicated
sessions exist in a same conference.

BREAKING CHANGE: * setCapatity
* setSpanForBringIn