fix(ConferenceCall): refresh the avatar url when merging
Because there is a bug of syncing between cache and store: when contact detail reducer getting
updated the data mapping does not get refreshed. So I have to trigger syncing by mannually call the
DataMatcher.prototype.match() while setting the ignoreCache=true.

BREAKING CHANGE: change the profile of participants into shape of
{avatarUrl:String,toUserName:String,partyNumber:String,rcId:String} to compat the inbound call
merging in the future

https://jira.ringcentral.com/browse/RCINT-8072